* BUGFIX don't write peakfile if the target directory is read only * REFACTORE OLTW code refactored and cleaned * IMPROVEMENT average mapping sequences that map to one and the same point in time to one single mapping

+++ AudioAlign/AudioAlign.Audio/Matching/Dixon2005/DTW.cs
@@ -225,6 +225,7 @@ namespace AudioAlign.Audio.Matching.Dixon2005 {
                 }
                 path.Add(new Pair(i, j));
             }
+
             return path;
         }
 
@@ -259,7 +260,7 @@ namespace AudioAlign.Audio.Matching.Dixon2005 {
             else if (weight > 1)
                 weight = 1;
 
-            return (int)(90d * d / sum * weight); // default scale = 90
+            return (90d * d / sum * weight); // default scale = 90
         }
 
         public static TimeSpan PositionToTimeSpan(long position) {
